pyproject.toml
[project]
name = "ai-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.3.0",
    "groq>=0.4.0",
    "gtts>=2.5.4",
    "markdown>=3.4.0",
    "mss>=6.1.0",
    "numpy>=1.20.0",
    "pillow>=9.0.0",
    "pygame>=2.6.1",
    "pynput>=1.7.6",
    "pyperclip>=1.9.0",
    "pyttsx3>=2.90",
    "pyyaml>=6.0",
    "requests>=2.28.0",
    "sounddevice>=0.4.5",
    "soundfile>=0.12.1",
]
